Use long for episode number

+++ TMDbLib/Objects/Credit/CreditEpisode.cs
@@ -18,7 +18,7 @@ public class CreditEpisode
     /// Gets or sets the episode number.
     /// </summary>
     [JsonProperty("episode_number")]
-    public int EpisodeNumber { get; set; }
+    public long EpisodeNumber { get; set; }
 
     /// <summary>
     /// Gets or sets the name of the episode.
