+++ TMDbLib/Objects/Search/TvSeasonEpisode.cs
@@ -27,7 +27,7 @@ public class TvSeasonEpisode
     /// Gets or sets the episode number.
     /// </summary>
     [JsonProperty("episode_number")]
-    public int EpisodeNumber { get; set; }
+    public long EpisodeNumber { get; set; }
 
     /// <summary>
     /// Gets or sets the episode type.
